+++ src/matyan/__init__.py
@@ -1,16 +1,18 @@
 from .utils import (
-    prepare_changelog,
-    json_changelog_cli,
+    generate_changelog,
     generate_changelog_cli,
+    json_changelog_cli,
+    prepare_changelog,
 )
 
 __title__ = 'matyan'
-__version__ = '0.2.1'
+__version__ = '0.3'
 __author__ = 'Artur Barseghyan'
 __copyright__ = '2019 Artur Barseghyan'
 __license__ = 'GPL-2.0-only OR LGPL-2.0-or-later'
 __all__ = (
-    'prepare_changelog',
-    'json_changelog_cli',
+    'generate_changelog',
     'generate_changelog_cli',
+    'json_changelog_cli',
+    'prepare_changelog',
 )